Bootstrap OWNERS files for the repository
In order to satisfy the blunderbuss and approvers plugins for the k8s CI
systems we are about to adopt, we need to add OWNERS files to
directories in Origin. The files contain reviewers and approvers for the
directory. Reviewers are members that blunderbuss will assign pull
requests to; approvers are members from whom an LGTM is necessary to
merge a pull request that touches files in the directory.

The inital values in these OWNERS files are generated by looking at the
most active contributors to any directory. The top ten most active
contributors to a directory are listed as reviewers; the top five most
active contributors that have merge rights with the current system are
listed as approvers.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>